+++ .gitignore
@@ -2,7 +2,7 @@
 .gradle
 build/
 !gradle/wrapper/gradle-wrapper.jar
-!**/src/main/**/build/
+!**/src/Main/**/build/
 !**/src/test/**/build/
 
 ### IntelliJ IDEA ###
@@ -14,7 +14,7 @@ build/
 *.iml
 *.ipr
 out/
-!**/src/main/**/out/
+!**/src/Main/**/out/
 !**/src/test/**/out/
 
 ### Eclipse ###
@@ -26,7 +26,7 @@ out/
 .springBeans
 .sts4-cache
 bin/
-!**/src/main/**/bin/
+!**/src/Main/**/bin/
 !**/src/test/**/bin/
 
 ### NetBeans ###
