+++ src/main/java/Main.java
@@ -1,0 +1,2 @@
+public class Main {
+}
