--- src/main/java/main.java
@@ -1,2 +1,0 @@
-public class main {
-}
